Collapse Materializer to thin shell; finish responder extraction
- responders.user.plain_preview_of_user; archives module for all archive schemas
- misc: article_column subscription + task schema
- can_read_answer in user_permission; search hit in responders.question
- materialize.py is pure re-exports + Materializer dispatch (~334 LOC)
- endpoints/services import responders or user_permission directly
