+++ chafan_core/app/api/api_v1/endpoints/answers.py
@@ -11,7 +11,7 @@ from chafan_core.app.services import answers as answers_service
 from chafan_core.app.common import OperationType, client_ip
 from chafan_core.app.endpoint_utils import check_writing_session
 from chafan_core.app.limiter import limiter
-from chafan_core.app.materialize import answer_archive_schema_from_orm
+from chafan_core.app.responders.archives import answer_archive_schema_from_orm
 from chafan_core.app.user_permission import check_user_in_site
 from chafan_core.app.schemas.answer import AnswerModUpdate
 from chafan_core.app.schemas.event import EventInternal, UpvoteAnswerInternal
