+++ chafan_core/app/responders/__init__.py
@@ -1,4 +1,5 @@
 from . import (
+    archives,
     question,
     submission,
     article,
